ErrorHandler func to class with static methods

+++ backend/src/lib/init/database.ts
@@ -2,6 +2,8 @@ import fs from 'fs'
 import config from '@config/DatabaseConfig.json' assert {type: "json"}
 import iniConfig from '@config/InitConfig.json' assert {type: "json"}
 import { Log, LogType } from '@lib/logging/log.js'
+import { Database } from '@lib/database/database.js'
+import {ErrorHandler} from '@lib/errorhandler/ErrorHandler.js'
 
 export const doesFileExist = (location: string, name: string) => {
     return fs.existsSync(location + name)
@@ -13,14 +15,25 @@ export const createFile    = async (location: string, name: string) => {
     fs.writeFileSync(location + name, '')
 }
 
-export const initialiseDatabase = (location: string) => {
+export const initialiseDatabase = async (location: string) => {
 
-    if(iniConfig.PERSIST_DB) return
-
-    if(doesFileExist(location, config.DATABASE_NAME)){
+    if(!iniConfig.PERSIST_DB && doesFileExist(location, config.DATABASE_NAME) ){
         Log(`Database file already exists, deleting...`, LogType.INI)
         deleteFile(location, config.DATABASE_NAME).then(() => Log(`Database file deleted: ${location}`, LogType.INI))
     }
 
     createFile(location, config.DATABASE_NAME).then(() => Log(`Created database file: ${location}`, LogType.INI))
+
+    try {
+        Database.setInstance(location + config.DATABASE_NAME).then(() => {Log(`Database instance set...`, LogType.INI)})
+
+        Log(`Testing database connection`, LogType.INI)
+        ;(await Database.getInstance()).authenticate().then(() =>
+            Log(`Database connection test successful`, LogType.INI))
+
+    } catch (err: unknown) {
+        let msg = "Unknown error occured"
+        if(err instanceof Error) msg = err.message
+        ErrorHandler.Error(msg)
+    }
 }
